feat: tweak assign batch/chunk logic

+++ coordinator/internal/types/get_task.go
@@ -3,7 +3,7 @@ package types
 // GetTaskParameter for ProverTasks request parameter
 type GetTaskParameter struct {
 	HardForkName string `form:"hard_fork_name" json:"hard_fork_name"`
-	ProverHeight int    `form:"prover_height" json:"prover_height"`
+	ProverHeight uint64 `form:"prover_height" json:"prover_height"`
 	TaskType     int    `form:"task_type" json:"task_type"`
 	VK           string `form:"vk" json:"vk"`
 }
